[Refactor] #495 - MyPagePushNotificationViewModel 삭제 및 MyPagePushNotificationViewContoller 수정
